Replace quickjspp, CMake
c-smile/quickjspp -> frida/quickjs
CMake -> Meson
