+++ .gitignore
@@ -1,6 +1,3 @@
 .vscode
 .idea
-build
-cmake-build-debug
-cmake-build-release
-src/psvmjs.c
+builddir
