--- src/psvmjs/.prettierignore
@@ -1,3 +1,0 @@
-# Ignore artifacts:
-build
-coverage
