--- src/psvmjs/.prettierrc
@@ -1,1 +1,0 @@
-{}
