--- src/psvmjs/package.json
@@ -1,30 +1,0 @@
-{
-  "name": "psvmjs",
-  "version": "1.0.0",
-  "description": "",
-  "main": "src/main.ts",
-  "type": "module",
-  "scripts": {
-    "build": "node ./scripts/build.js",
-    "lint": "eslint ./src/** --ext .ts --report-unused-disable-directives --max-warnings 0"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "MIT",
-  "devDependencies": {
-    "@sprout2000/esbuild-copy-plugin": "^1.1.13",
-    "@typescript-eslint/eslint-plugin": "^6.21.0",
-    "@typescript-eslint/parser": "^6.21.0",
-    "esbuild": "^0.20.0",
-    "eslint": "^8.56.0",
-    "eslint-config-prettier": "^9.1.0",
-    "prettier": "3.2.5"
-  },
-  "dependencies": {
-    "@pkmn/randoms": "^0.8.7",
-    "@pkmn/sim": "^0.8.7",
-    "@pkmn/streams": "^1.1.0",
-    "nanoid": "^5.0.5",
-    "polyfill-crypto.getrandomvalues": "^1.0.0"
-  }
-}
