--- src/psvmjs/src/globalize.js
@@ -1,4 +1,0 @@
-// noinspection JSFileReferences
-import * as psvm from "./psvm.js";
-
-globalThis.ShowdownService = psvm.ShowdownService;
