+++ subprojects/psvmjs/.prettierrc
@@ -1,0 +1,1 @@
+{}
