+++ subprojects/psvmjs/src/globalize.js
@@ -1,0 +1,4 @@
+// noinspection JSFileReferences
+import * as psvm from "./psvm.js";
+
+globalThis.ShowdownService = psvm.ShowdownService;
